third_party/blink/web_tests/external/wpt/html/browsers/browsing-the-web/history-traversal/same-url-expected.txt

This is a testharness.js-based test.
FAIL Test same-URL navigation and its effects on history assert_equals: expected 3 but got 2
Harness: the test ran to completion.
